+++ Gateway/Request/AuthorizationRequest.php
@@ -186,6 +186,11 @@ class AuthorizationRequest implements BuilderInterface
         $values['PBX_SHOPPINGCART'] = $payment->getXmlShoppingCartInformation($order);
         $values['PBX_BILLING'] = $payment->getXmlBillingInformation($order);
 
+        // Check for 3DS exemption
+        if ($payment->getConfigData('exemption_3ds_max_amount') && $orderAmount <= $payment->getConfigData('exemption_3ds_max_amount')) {
+            $values['PBX_SOUHAITAUTHENT'] = '02';
+        }
+
         // Sort parameters for simpler debug
         ksort($values);