+++ Model/Payment/Threetime.php
@@ -28,19 +28,11 @@ use \Magento\Framework\Validator\Exception;
 class Threetime extends AbstractPayment
 {
     const CODE = 'pbxep_threetime';
-    const XML_PATH = 'payment/pbxep_threetime/cctypes';
 
     protected $_code = self::CODE;
-    protected $_hasCctypes = true;
     protected $_allowManualDebit = true;
     protected $_allowDeferredDebit = true;
     protected $_allowRefund = true;
-
-    public function getReceipentEmail()
-    {
-        $storeScope = \Magento\Store\Model\ScopeInterface::SCOPE_STORE;
-        return $this->_scopeConfig->getValue(self::XML_PATH, $storeScope);
-    }
 
     public function toOptionArray()
     {