+++ Model/Ui/PbxeppaybuttonsConfigProvider.php
@@ -60,6 +60,11 @@ final class PbxeppaybuttonsConfigProvider implements ConfigProviderInterface
 
         $cards = [];
         $types = $pbxeppaybuttonsConfig->getCards();
+
+        if (is_null($types)) {
+            return $cards;
+        }
+
         if (!is_array($types)) {
             $types = explode(',', $types);
         }