+++ Model/Ui/PbxepprepaidConfig.php
@@ -56,12 +56,6 @@ class PbxepprepaidConfig
         $this->scopeConfig = $scopeConfig;
     }
 
-    public function getReceipentEmail()
-    {
-        $storeScope = \Magento\Store\Model\ScopeInterface::SCOPE_STORE;
-        return $this->scopeConfig->getValue(self::PAYMENT_METHOD_PBXEPPREPAID_XML_PATH, $storeScope);
-    }
-
     /**
      * @return string
      */