+++ Model/Ui/PbxepthreetimeConfigProvider.php
@@ -60,6 +60,11 @@ final class PbxepthreetimeConfigProvider implements ConfigProviderInterface
 
         $cards = [];
         $types = $pbxepthreetimeConfig->getCards();
+
+        if (is_null($types)) {
+            return $cards;
+        }
+
         if (!is_array($types)) {
             $types = explode(',', $types);
         }